requirements.txt
asgiref==3.9.2
Django==5.2.6
django-environ==0.12.0
psycopg==3.2.10
psycopg-binary==3.2.10
sqlparse==0.5.3
tzdata==2025.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3